+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -44,5 +44,5 @@ namespace llvm{
 	class CallInst;
 	class LoadInst;
 	class GlobalVariable;
-	
+
 }