+++ PCIT_core/internal/llvm_interface/include/llvm_interface.h
@@ -20,4 +20,5 @@
 
 #include "./IRBuilder.h"
 #include "./ExecutionEngine.h"
+#include "./OrcJIT.h"